Reduce kubelet LOC: extract getters
